Widen the clickable area for statuses in grouped notifications
Fixes #31089
